refactor: rename 'supported' → 'synchronous' query, move helper stuff
